organize code to isolate solidity mirror code
